refactored all geoips output_checkers to be class based
