+++ geoips/interfaces/__init__.py
@@ -17,11 +17,11 @@
 
 from geoips.interfaces.class_based.algorithms import algorithms
 from geoips.interfaces.class_based.colormappers import colormappers
-from geoips.interfaces.class_based.output_checkers import output_checkers
 from geoips.interfaces.class_based.coverage_checkers import coverage_checkers
 from geoips.interfaces.class_based.databases import databases
 from geoips.interfaces.class_based.filename_formatters import filename_formatters
 from geoips.interfaces.class_based.interpolators import interpolators
+from geoips.interfaces.class_based.output_checkers import output_checkers
 from geoips.interfaces.class_based.output_formatters import (
     output_formatters,
 )
