+++ geoips/interfaces/class_based_plugin.py
@@ -81,6 +81,8 @@ class BaseClassPlugin(ABC):
     developers to implement this functionality in every plugin class. Initially, this
     will be used to convert inputs from DataTree to other formats and back to DataTree
     after processing, but it could be used for other common tasks as well.
+
+    NOTE: add an init function that takes module as an argument [optional]
     """
 
     # If set to True, we are in OBP. False means we are in a legacy procflow.
@@ -196,20 +198,8 @@ class BaseClassPlugin(ABC):
             if attribute_checker is not None:
                 attribute_checker(cls)
 
-        cls_registry_entry = getattr(
-            interfaces, cls.interface
-        ).plugin_registry.registered_class_based_plugins[cls.interface][cls.name]
-
-        true_class = True
-
-        if cls_registry_entry["is_derived_plugin_object"]:
-            # Temporarily ignore 'classes' that are still derived from modules. We'll
-            # remove this when we completely convert over to class-based plugins
-            true_class = False
-            cls.call = cls.__dict__["__call__"]
-
         # Prevent overriding __call__ in a True class-based plugin
-        if "__call__" in cls.__dict__ and true_class:
+        if "__call__" in cls.__dict__:
             raise TypeError(f"{cls.__name__} cannot override __call__")
 
         try:
